\name{bayesdnamix-package}
\docType{package}
\title{
Procedure for evaluate marginal likelihood for quantitative STR DNA mixture models}
\description{
bayesdnamix contains a procedure for evaluating marginal likelihoods for quantitative bayesian models for STR DNA mixtures for general specifiactions of hypothesis.}
\details{
\tabular{ll}{
Package: \tab bayesdnamix\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-10\cr
License: \tab  GPL(>=2) \cr
}
}
\author{
Oyvind Bleka <Oyvind.bleka@fhi.no>
Maintainer: Oyvind Bleka <Oyvind.bleka@fhi.no>
}

