% Generated by roxygen2 (4.0.1): do not edit by hand
\name{contLikDrop}
\alias{contLikDrop}
\title{contLikDrop}
\usage{
contLikDrop(nC, mixData, popFreq, refData = NULL, condOrder = NULL,
  prD = NULL, prC = 0, model = 1, pTau = function(x) {     return(1) },
  taumax = 100, maxeval = 10000, threshT = 50, fst = 0)
}
\arguments{
\item{nC}{Number of contributors in model.}

\item{mixData}{Evidence object with list elements adata[[i]] and hdata[[i]]. Each element has a loci-list with list-element 'i' storing qualitative data in 'adata' and quantitative data in 'hdata'.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{refData}{Reference objects with list element [[s]]$adata[[i]]. The list element has reference-list with list-element 's' having a loci-list adata with list-element 'i storing qualitative data.}

\item{condOrder}{Specify conditioning references from refData (must be consistent order). For instance condOrder=(0,2,1,0) means that we restrict the model such that Ref2 and Ref3 are respectively conditioned as 2. contributor and 1. contributor in the model. condOrder=-1 means the reference is known-non contributor!}

\item{prD}{A vector of allele drop-out probabilities (p_hom=p_het^2) for each contributors. Default is 0.}

\item{prC}{A numeric for allele drop-in probability. Default is 0.}

\item{model}{A integer for specification of model. See details for more information.}

\item{pTau}{Prior function for tau-parameter. Flat prior is default.}

\item{taumax}{Maximum range of tau-parameter. Default is 100.}

\item{maxeval}{Maxumum number of evaluations in the interale function.}

\item{threshT}{The analytical threshold given. Used when considering allele drop-outs.}

\item{fst}{is the coancestry coeffecient. Default is 0.}
}
\value{
lik Marginalized likelihood of the hypothesis (model) given observed evidence.
}
\description{
contLikDrop evaluates the marginal likelihood of the STR DNA mixture given some assumed model by integrate out parameters for given allele drop-out probability for each contributors.
}
\details{
The procedure are doing numerical integration to approximate the marginal probability by integrate over noisance parameters. Mixture proportions have flat prior.

The user may specify probability of drop-out for each contributors.

The peak heights are scaled between [0,1] such that the peak heights are not accounted for in the models.

Function calls procedure in c++ by using the package Armadillo

Models: model={'0'-unit weights,'1'-mixsep}
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}
\keyword{Bayesian}
\keyword{continuous,}
\keyword{models}

